requirements.txt
cython==0.29.28
Pillow-SIMD==8.4.0.post0